Better token management, better test suite and flow control parsing
Renamed Token to Tokens and factored out the token deque to the Tokens
class. Added a better example test code as the test_suite.cl which
should provide coverage for all of the language features. Started flow
control structures for If, Else, For and While, For does not work yet